requirements.txt
openai==0.27.0
python-dotenv==1.0.0
Flask==2.2.3
Jinja2==3.1.2
pyyaml==6.0
psycopg2==2.9.6